Extract cross-world playthrough traversal into shared header
Move WriteComboPlaythrough's sphere-collect body into combo/rando/ComboPlaythrough.h
(ComboRando::RunPlaythrough, MM-restore passed in, returns a result struct); ComboShip.cpp
now delegates to it. Lets the headless validator reuse the exact same traversal.
